Upgrade Nomad and Consul
This introduces the unstable nix channel to fix some issues with the
Nomad install. See, the Consul Connect sidecar defaults to an old docker
tag without ARM support, causing deploy failures at a ratio of 2:3. Or
at least, it fails because of that. It's still failing for other reasons
yet to be determined.