Force downgrade Envoy (again)
I upgraded everything in the datacenter and thought surely the Envoy
image runs on Raspberry Pi now. They explicitly added support for ARM!
It seemed like it was on the cusp of support!

Turns out, no. And I don't see them adding it in the future either. It
may be time to investigate alternatives to the Envoy sidecar.

See: a7e3939