Add service for running a router on NixOS
I've been using OPNsense for a while as my home router, but the
configuration is all mouse driven and much of the product is just
a complex UI over features that are already built into Linux. I'd much
rather manage it declaratively the same as the rest of my
infrastructure, then I can leverage the full power of the dark sid...
er, NixOS.

Right now I'm able to access my router at `selfhosted.city`, but only
through the LAN interface. I need to move control of that domain to the
new private ingress server, and ideally run ACME near a Vault service
instead. The goal is to free up the domain and enable tighter network
controls, like VLANs, before opening anything up to the internet.