Use one of the RPi 3s as a router
I'm currently using the Pi as a router. This makes me extraordinarily
happy. I'm having a bit of trouble with DNS, which only seems to be
resolving on the router but hanging on downstream requests. I haven't
figured out why, so for the meantime I'm just disabling the service.
I may replace it with an alternative.