Restrict firewall holes to specific interfaces
I didn't realize you could open holes in the firewall on a per-interface
basis. Now DNS and SSH are completely hidden via the firewall. This
shouldn't matter in practice, as DNS is only bound to local interfaces
and SSH authorizes by public key, but it's reassuring to have the
firewall as added protection.