Reinstall NixOS on the former OPNsense server
I've completely migrated off of OPNsense and onto a custom-built
raspberry pi. Now the old desktop can be part of the compute cluster
again. That computer's potential was wasted before.

I also changed that computer's registered MAC address to use the wifi
hardware instead of ethernet. I keep it in a corner separate from the
rack because it doesn't really fit.