Avoid entering encryption keys over SSH
There are serious downsides. For one, the SSH keys have to be stored
insecurely on the root partition. Anyone with physical access can swipe
the keys and MITM the decryption. If I'm gonna do this right, and I aim
to, I should enter the decryption keys manually. I might even use this
service to auto-unseal Vault. Security is very important here.