Colocate IP assignments with host configs
This is practical for two reasons: first, I no longer have to specify
the host name, as it's inferred by `networking.hostName`. Second, `node`
lookups have access to a static IP. This is important because I can
filter by all nodes running Consul, grab their IP address, and generate
`A` records that load balance over DNS.

The only downside, and one which made me seriously consider doing it
differently, is it's no longer clear which IP addresses are reserved.

I plan to generate the assignments in the future.