Split out Consul server and client roles
I've split out the Consul server agent into an opt-in service. This is
useful for noisier machines that I don't always keep on. I need them to
work as a cluster when they're enabled, but not break the quorum on
downtime.

The Consul servers are accessible through the `consul.service.{domain}`
record.