Enable Nomad on Multivac
Multivac isn't online all that often, but when it is, I'd like it to
pull its fair share. Nomad self-heals well enough to detect when I shut
it down (and I can improve the drain process if need be).

This is partly why I split out the server role. Shutting down Multivac
before would cripple consensus.