Proxy DNS service lookups to Consul
Instead of manually adding these records for every service, I'm
forwarding all DNS queries to `service.{domain}` over to Consul, which
is configured to answer under that zone.

The main advantage is health checks. If a nomad server goes down, it's
automatically removed from DNS queries and the Nomad UI will continue to
work reliably.

The second advantage is discovery of Nomad services. I just used this to
map an apex record to all the private ingresses. It doesn't work yet,
I haven't figured out Consul Connect Gateways, but the DNS works! That's
a step in the right direction.