Use TLS on Consul endpoints
The http endpoint is only accessible on localhost now, and TLS
certificates are provisioned for all other services. RPC calls are still
unencrypted because that requires a token, and I can use Vault for that.
I'll do that later.

I've got this written down elsewhere, but it's an important note: for
the `verify_server_hostname` feature to work, the Vault certificate must
work with `server.lab.selfhosted.city`. I'm using a SAN.

I haven't gotten HTTPS to work outside the cluster, even manually
accepting the certificate. Bits in the documentation suggest it might be
trying to validate my browser's certificate too, which obviously won't
work. I'm using SSH tunneling in the meantime.