Restrict key access to their respective services
Services must be added to the `keys` group to list keys or read the
directory, and each key is only readable to the corresponding user.

I'm not confident that Nomad won't go shifting GID/UIDs given the way
the service is defined. This might break in the future. I'll keep an eye
on it.