Extract admin config into service file
The entire admin role and all it's associated tools have been extracted
into a service file. Now it can be disabled optionally and paves the way
for configuration. I didn't like it hanging out in the host definition,
it was a fair amount of code with a related theme that didn't need to be
together.

I've also changed the default shell back to bash for all other users.
Now root doesn't have an error trying to load starship. I could've added
it to the path, but I like the distinctive prompt that bash has for
root.