Store Nomad certs in a consistent place
I've changed my opinion on storing certificates on disk recently. This
way computers can reboot without immediately requiring keys - the
certificates will coast for a while. This is hugely important for
backbone services like Consul and Vault. If I have to unplug the
datacenter, I don't want to manually recover it because the certs got
lost and nothing can be reissued.

This is less of a concern for Nomad, but I think it's better to keep it
consistent.