Run Nomad clients with higher privileges
HashiCorp recommends running Nomad clients under root and servers under
the `nomad` group. Some services were failing because Nomad couldn't
update the networking rules, which makes perfect sense if permissions
are dropped. It only seems to happen if it starts a new service, so
I suspect that I had this set a while ago and it got killed in the
refactors, but I didn't notice because Docker containers persist between
NixOps deploys.