Fix TLS in Nomad -> Consul gRPC
All Nomad services were failing to register with the local Consul
daemon. First, it was using the `http` port when it should've been
encrypting, second I didn't add any certificate configurations for the
Envoy sidecar so it was attempting an unencrypted socket and failing
with other errors. I've forced TLS on the Consul connection.

A challenge in testing all this was my out-of-date Nomad/Consul versions
(held back by Envoy's support for raspberry pi) and the slowness of
testing new features as Nomad can't dynamically update certificates for
the sidecar and has to go through the whole ordeal of rescheduling
everything.