Add support for multiple subnets
The goal here is to isolate my network into distinct chunks, like IoT
devices, guest networks, trusted computers, and the datacenter. Each
subnet may have different routing rules.

This commit replaces the concept of a "lan" interface with several
subnets, each defining their own link/interface. This is a somewhat
clunky idea, it isn't a perfectly accurate binding, but it's close
enough for my use case.

The next step is to link my Unifi WiFi access point directly to the
router and start extending it with vlans and dedicated subnets.