Track DNS blocklist as flake input
This used to be inlined as a `fetchurl` call. It was tedious to update
because it was pinned to a tag and hashed, and every time it fell out of
date I needed to change the tag and regenerate the hash, manually
changing the source code.

I looked for a good way to pass this down the call stack and eventually
noticed `_module.args` during an upgrade to 22.05, which functions
identically to `specialArgs` in the NixOS builder.

It solves my problem and so I brought it up as a potential solution in
this related NixOps issue: https://github.com/NixOS/nixops/issues/1494

Turns out, I'm a n00b. It's been there since around 2015. I can use it
even on my outdated version.

This is the new update procedure for the DNS blocklist:

```sh
nix flake lock --update-input dns-blocklist
```

Much nicer!