Disable all non-critical lab services
In the process of moving to New York, my Vault certificates expired and
the house of cards collapsed. Further, my ISP is stingy with IPv4
addresses and I need some serious upgrades to my router.

I've learned a lot about NixOS over the past several years. As scary as
it is, I think it's time to start fresh. A snapshot of my old
configuration is available on the branch named "ye-olden-days".

I'll keep the old service files as a reference, but for now, everything
except my router is disabled. I'll gradually rewrite stuff and turn them
back on.