Prune and modernize home lab
I've learned a lot the last few years of working with NixOS. Some
patterns in my home lab were not very good. I didn't know how to design
modules, organize hosts, or manage secrets very well. I still don't have
all the answers, but I've learned a lot.

It's been hard to find the motivation to work on the home lab when
there's so much that has to be fixed. Rewriting is rarely the right
solution, but this is a hobby project. It's main value is bringing joy.
And if rewriting it to a better architecture brings me joy, then let's
give it a go.

Most of the old modules were dead. The vault service collapsed when
I moved to New York because the home lab was offline for long enough
that the certificates expired, bringing the rest of the cluster with it.
Some services also got deprecated and with my new job, I didn't have
time to migrate, so everything kind of rotted.

This commit removes a lot of the "dead code", even though it wasn't
technically dead. I'll try it again, but hopefully with a more robust
design that doesn't require so much attention. (Or at least one that's
easier to rebuild from scratch.)