Defer pool decryption until after boot
My file server was kind of broken because I set everything up to
automount the ZFS datasets, but I forgot that the pool needed manual
decryption so the next time I cycled power, the rpi got stuck at stage
1 boot. I couldn't even ssh into the machine.

I tried a few ways around it, specifically trying to add sshd to the
boot stage, but limitations on initrd secrets management prevented me
from doing it securely. I'd essentially have to make my SSH keys
world-readable. This wouldn't be an issue with a typical boot loader
like grub.

But whatever the case, I don't like the idea of it blocking my boot
sequence anyway, so I deferred decryption to a script and systemd
target. I tried to put the script in a systemd oneshot, but got hung up
on linking it with the tty so I eventually split it into a target and
a script pair. It works well enough.

I learned more about systemd than I wanted to today.

Note: this is admittedly a hack, but the file server isn't a load
bearing structure in my home lab. It's used for backups, but it won't
take any services down if it drops. I don't want to invest time in
making it generic yet.