Assign names to vlans/eth devices
Now instead of `vlan2@eth1` I have `vlan-guest@wap`. This helps decouple
the hardware from the router configuration. I'm hoping to move this from
my rpi3 to an rpi4 and that requires a little more abstraction. It's
also quite hard to figure remember what `vlan2` means without the
context of comments, so this is a QOL improvement too.