Upgrade NixOS to 23.05
The 23.11 release is just around the corner and I hadn't even upgraded
to NixOS 23.05. This upgrade was a little tricky because it seems
extlinux boot loaders are sized differently and caused corrupted offsets
on rpi3 devices. I had to reinstall from the latest hydra image and
bootstrap from there. Thankfully it worked out of the box, it was just
an issue of migrating an existing system. Now that the router is on an
rpi4 I'm free to nuke those older ones and start from scratch.