Rename servers to reflect hardware class
The form factor holding my home lab is a stack. Whenever I add another
device, it goes on top of the stack. I can never remember what codename
corresponds to which device. It's challenging every time I need to
attach a KVM setup. Having a numbering scheme and a hardware class to go
by would greatly simplify administration. Which device is rpi4-002? It's
the second rpi4 from the bottom. Easy.