Tidy router config
This fixes a number of things that have been bothering me. For one, the
association of `udev` rules and the chosen network interfaces were
uncontrolled. Now changing one changes the other.

This also fixes a deprecation warning on one of the Avahi mDNS options.

I've also shuffled a few options around to look nicer.