Add Nix tools to manage ZFS pools and datasets
These new options generate scripts in the task runner to initialize and
manage ZFS pools and datasets, including properties.

This was kind of tedious and I'm not thrilled with the result. I'm
viewing it as exploratory work that I hope to change in the future.
Moving management of pool/dataset properties to a python script that can
be applied even after things are initialized would be very useful.