Define unified administration command
This (finally!) leverages my `clapfile` experiment to provide a unified
CLI for managing the system. The first consumer is the file-storage
service, exposing a few subcommands.

Any subcommand can be added via `lab.system.${subcommand}`. The
configuration is defined by `clapfile`.

I ran into a painful speed bump during testing. It took me a while to
realize that the node package set was not inheriting overlays from the
host. It might've been more obvious on a different arch. Anyway, now the
overlays are passed through. This adds `pkgs.clapfile` to test
environments as well as the `pkgs.unstable` package set.