Add nushell command to sync zprops
This is a rewrite of the Python implementation. I am seriously
disappointed with the Python outcome. It's large, tedious, and somewhat
horrifying. I wouldn't merge it because it's not worth the cost.

While building it in Python, I was using Nushell to test high-level
ideas of the program. About halfway through it occurred to me that
Nushell is probably the better place to implement this. It has excellent
support for command output parsing, data transformations (diffs,
execution plans), running commands (obviously), and printing nice output
(`ansi` module).

I could do it faster, easier, better, and with less code by writing it
in Nushell.

Or so goes the theory.

This is a hobby project, so why not try.