Use subflake for Nushell dev environment
This replaces the `nix-shell` approach with a flake. The advantage is
ensuring I'm testing against the same version of Nushell that gets
deployed. It also means I can extend my environment with scripts rather
than replacing it, so the devx is seamless.

The drawback is that subflakes aren't well supported. I had to use
a userspace implementation of `builtins.getFlake` to avoid pinning the
parent flake in the child's lockfile. It's already implicitly locked by
git and the noise of a constantly updating lockfile is unacceptable (not
to mention undermining the whole point, which is ensuring tests run with
the same versions as the parent).

Hopefully better patterns emerge in the future.