Move ZFS control under `lab.filesystems.zfs`
I'm experimenting with `zfs` and `bcachefs` as root FS on my RPI
machines. This is more attractive now that I have control over advanced
properties. (Previously it would've been a bunch of undocumented setup.)

If I use the ZFS modules without encryption or running a file server,
it's more a filesystem module than anything else. I've renamed it to
match that model.