Use Nushell as default shell
I use this zealously on all my other machines. I hadn't put it here
because there was always something more important, and `nixpkgs` was out
of date for a long time. Now it's ready, and the lab is in a space where
I can add home-manager. This brings a lot more configuration power
through the HM framework.

This is still bare-bones.