Add sandbox test environment
I've been using test VMs a lot more, recently. It's a great way to
experiment with new services and programs. For example I just used a VM
yesterday to get familiar with `bcachefs` now that it merged into Linux.
I never would've done that on my home lab or personal machine as it
required upgrading the kernel and screwing with the file system. Doing
it in a VM also means I avoid the remote Nix builders, which can be dog
slow because I'm farming to Raspberry Pi machines.

This is a much faster, safer way to experiment.