Create DNS lab service
This is extracted from the router module, but with significant changes
and more options. I've also added tests to cover the biggest features.

My goal is to break apart the router into smaller services, similar to
what I did with the DHCP module. By the end, the router should be
a small glue layer binding DNS, DHCP, NAT, and firewalls.