Replace nested flake with specialized dev shell
The ZFS development environment was using a pretty sketchy `call-flake`
pattern in order to extend the parent flake's development shell. Now I'm
generating a set of development shells, all extending the base shell,
using a customization function specified in the project.

It's still super weird, but it keeps me on the well-supported path of
Nix. The old way had some risk of breaking everything if Nix upgrades
and no longer allowed the hacks used by `call-flake`.