Add markdown and manpage doc generators
I'm not installing these anywhere yet, but it paves the way. It took
a bit to figure out how this worked. I've found several ways not to do
it, and this is the first implementation that seemed semi-stable.

The dream is to plug this into a static site built on every push.