Remove cfg link to `lab.networks` for DHCP server
In the module system I'm using an aliased option into `lab.networks.*`
for convenient access. It's caused a number of problems, especially with
derived state like `ipv4.gateway`. Those should be read-only, but
copying the value onto another part of the module system means it's
getting set from an external source. I had to open permissions to allow
that.

It's a little gross.

This commit is a step away from that pattern. I need to do the same in
the gateway module. `lab.networks` should be the sole source of truth.