Use single set of IP addresses for DNS
It used to be that only the gateway ran the DNS server, which allowed me
to listen on all subnets and expose a local IP. That's still the case,
but I'm working towards HA mode for bedrock services. In the near future
DNS servers may be running on machines that aren't joined to every
subnet (not on gateways).

In effect, this means I can't make DNS IP addresses local to the subnet.
I'm not going to run a dedicated DNS server on the guest subnet, for
example. Only the gateway exists on all subnets.

So instead, I'm deriving a set of all IPs that are running DNS
(according to `host.ip4`) and setting that as the nameservers list. This
might degrade network performance because clients need to route through
the gateway to reach another subnet and it introduces a SPOF, but no
more than I already had.

This may prevent me from enforcing firewall rules between subnets.
I might need a subnet just for bedrock services with more permissive
access rules. I've already considered this for other reasons. TBD.