Use standard test harness from nixpkgs
Previously I was importing a file by path inside the nixpkgs tarball to
get access to the Module-ified `nixosTest` harness. Apparently it was
already exposed via `testers.runNixOSTest`. I discovered this through
this article:
https://nixcademy.com/2023/12/01/nixos-integration-tests-part-2/

The ergonomics are slightly better. For example, it uses `pkgsLinux`
instead of re-evaluating a new fixpoint, and this means my overlays
automatically inherit. Neat.