Prove out multi-tenancy with nspawn/nixos
This works. I've been trying to prove the concept for a long time, but
never got it working. The goal is for each container to have a private
network and listen on DHCP for a unique address. Just about all services
would live in these containers.

That buys me a few things: no conflicts between services (ports, files),
stronger security (can't bypass auth proxies by hitting localhost),
multiple instances of the same `services.*`, and keeping a traditional
maintenance interface (ssh, NixOS modules).

The caveat with macvlans (and the reason it took me so long to discover
a working implementation) is it doesn't work over wifi. The containers
have to be connected over ethernet. And all my tests were from my
laptop...