Add support for DHCP reservations by client ID
In addition to mac address associations, now I can provision IPs by DHCP
client ID associations. The advantage of decoupling IPs from mac
addresses is it opens the possibility of macvlans which are more
difficult to manage (it's tedious to spoof a macvlan address).

I'm currently pursuing macvlans as a solution for NixOS container
multi-tenancy where they obtain IP addresses through DHCP in a private
network namespace. I don't plan to manage those addresses manually, but
due to the way macvlans are networked, the hosts must also use macvlans
which means I can't rely on a stable mac address.

Client identifiers are generated by the IP address because I don't care
what the ID is, only that it's stable and corresponds to my IP address
record. IP addresses are a simple way to achieve it. I could've also
used the machine's hostname, but Kea doesn't seem to appreciate non-hex
value associations and crashes on startup. Formatting the IP to a hex
value solves that.

I also prefixed those values with `FE:01` juuuust in case it was being
parsed as a real mac address. I don't want conflicts on any real
hardware. From what I could tell, `FE:01` has not been allocated and may
never be allocated (it wasn't clear if it was reserved or not, didn't
care enough to be certain).