Put all hosts behind macvlan interfaces
Well, all except the router. That host requires a bit more thinking.

This commit makes it so instead of requesting a DHCP address for `eth0`
or whatever the primary interface is, it creates a macvlan and requests
it there instead.

The advantage is containers. I can create containers with private
networking that get their own IP automatically assigned. This is nice
because I can ssh directly into them and route to them like any other
host on the network. The drawback is hosts can't communicate with
containers or vice versa unless they're also using a macvlan. It sees
the target as a different mac address and defaults out to the switch,
which drops the packet because the destination would've reflected right
back and obviously that can't be right (unless it's VEPA-friendly).

So the solution is putting the primary interface behind another macvlan.
That means outbound traffic destined for a mac address in the local
macvlan table doesn't need to touch the lower interface and always
routes to the correct recipient. If it's not a local macvlan address, it
continues to the network as usual. This avoids dropped packets on
reflection.