Add proof of concept container tools
The goal is to have most framework stratum services running in
containers. This helps them play nicer with port and resource allocation
(a lesson from my last home lab) and also helps the security posture as
I can limit access to known ports and install automatic proxies with
mTLS. But that will come later. The first step is making this work
reliably.