Use etcd for DNS service discovery
It's not hooked up to any live data sources yet. That's my next step.
The goal is to replace these DHCP client IDs with dynamic hostname
associations that happen on lease renewal. Sitting in this middleground
of dynamic and static allocations is so painful. One of my hosts is
getting completely ignored by the reservation pool and it happened after
deploying a container which is suspicious (macvlan trickery?)

I'd rather abolish the whole reservation idea and allocate subnets by
role instead of specific IP.