Add support for static etcd records
This is a little bit of a kludge, but it's good enough. It allows me to
query the router on the LAN which would've been difficult because it's
a static IP and doesn't have a DHCP lease, which means it isn't
automatically added to the database.

Now I'm adding the record as an etcd post-start script. Always
guaranteed to be there.

The risk is that I break DNS, or leave a stale record that I forget to
update.