Expose VPN through a CF tunnel
This is an experiment. I want to share my VPN with a fellow home-lab
operator with restricted ACLs to shared resources. Having a public VPN
to bootstrap the connection is useful, not only for that case, but
bootstrapping my own devices.

I'm using Cloudflare for TLS termination. I'm not thrilled that they
could MITM me, but the risk seems low and the reputational damage to
them would be high, so I'm not too worried about it. Plus, it would only
help if they MITM'd me during bootstrap. (I think.)