Fix tests in propctl
I fiddled with Disko hoping to replace this code. I came back
unsatisfied. Disko seems to be targeting the initial install and boot
image. It doesn't handle changes to an existing configuration, such as
changed filesystem properties or renamed pools. It also makes everything
significantly more difficult for remote deploys because it expects
a `configuration.nix` file which obviously won't exist on the remote
end.

So, I'm returning to my `propctl` approach. This fixes the tests and
minor broken syntax through nushell upgrades.